// apps/web/src/modules/vehicles/constants/vehicle.enums.ts
export enum VehicleType {
  BIKE = 'BIKE',
  CAR = 'CAR',
  TRUCK = 'TRUCK',
  EV = 'EV',
  COMMERCIAL = 'COMMERCIAL',
}

export enum FuelType {
  PETROL = 'PETROL',
  DIESEL = 'DIESEL',
  ELECTRIC = 'ELECTRIC',
  CNG = 'CNG',
  HYBRID = 'HYBRID',
}

export enum TransmissionType {
  MANUAL = 'MANUAL',
  AUTOMATIC = 'AUTOMATIC',
}